Make both linters search for lint_config file.
Both moonc and moonpick can read a lint_config file.
For moonc, a lint_config.lua can be in the current directory.
For moonpick, a lint_config.lua or lint_config.moon can be in
the current or a subdirectory.
The moonc checker needs to use source-inline, because the
lint_config applies configurations per filepath.